Work on ThingsDB

+++ inc/ti/tcp.h
@@ -6,7 +6,7 @@
 
 #include <uv.h>
 
-const char * ti_tcp_ip_support_str(uint8_t ip_support);
+const char * ti_tcp_ip_support_str(int ip_support);
 char * ti_tcp_name(uv_tcp_t * client);
 
 #endif /* TI_TCP_H_ */
